backend/requirements.txt
# FastAPI & server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
motor==3.3.2
pymongo==4.6.1

# ML / Vision / Audio
torch>=2.1.0
torchvision>=0.16.0
opencv-python-headless==4.9.0.80
numpy>=1.24.0
Pillow>=10.0.0
librosa>=0.10.1
soundfile>=0.12.1
scipy>=1.11.0

# Utils
python-dotenv==1.0.1
pydantic-settings==2.1.0
aiofiles==23.2.1
python-magic==0.4.27
